src: Fix axi_write_buffer
* Fix the `_d` to `_d` assignment is a typo
* If there is an ingress and an egress at the same time, the counter will
count down 1 instead of staying the same
* For if (egress_w_last & mgmt_ready_aw), in the case of a multi-beat
write, wouldn't the AW handshake long be finished when the last W beat
leaves the buffer? We saw the num_lasts counter not being decremented
and overflowing when preloading from the debug module using progbuf,
which caused a deadlock (monitored on the FPGA with ILAs). This was
tested by using the `axirt_budget` test from Cheshire as a means of
configuration and then trying to preload anything using GDB.